+++ backend/data/models/models.go
@@ -145,7 +145,7 @@ type Notice struct {
 	PartyID int
 	Type    string
 	Source  string
-	Data    *string
+	Data    []byte
 }
 
 type Notification struct {